Made background sun change
6am, between 6am-8pm , 8pm and nighttime

+++ WeatherApplicationAndrew/app/src/main/java/com/example/andrew/weatherapplicationandrew/Common/Common.kt
@@ -24,7 +24,7 @@ object Common {
         return sb.toString()
     }
 
-    fun unixTimeStampToDateTime(unixTimeStamp:Double): String{
+    fun unixTimeStampToDateTime(unixTimeStamp:Int): String{
         val dateFormat = SimpleDateFormat("HH:mm")
         val date = Date()
         date.time = unixTimeStamp.toLong()*1000
