+++ WeatherApplicationAndrew/app/src/main/java/com/example/andrew/weatherapplicationandrew/Model/Sys.kt
@@ -1,3 +1,3 @@
 package com.example.andrew.weatherapplicationandrew.Model
 
-class Sys(var message:Double,var country:String?,var sunrise:Double,var sunset:Double)
+class Sys(var message:Double,var country:String?,var sunrise:Int,var sunset:Int)
